Move filing service into the default-memory plugin lifecycle
The periodic PKB filing/compaction service is a memory-system concern, so
relocate it into the default-memory plugin and drive its start/stop from the
plugin's init/shutdown hooks instead of bespoke lifecycle/shutdown wiring:

- Move filing-service.ts under plugins/defaults/memory/ (FilingService class
  unchanged; the standalone startFilingService/stopFilingService wrappers are
  gone — their bodies are now the hooks).
- Add memory plugin hooks/init.ts (start, skipped under memory v2) and
  hooks/shutdown.ts (stop + drain), wired into defaultMemoryPlugin.hooks.
- Drop the startFilingService() call from daemon startup and the
  stopFilingService() call from the shutdown sequence; the plugin bootstrap
  now starts it during initializePlugins() and the plugin teardown stops it
  during runShutdownHooks("daemon-shutdown").
- Repoint filing-routes and the two filing tests at the new path.

A bootstrap test asserted no default plugin contributes an init hook; updated
it to assert the disabled plugin's hook specifically is filtered, since
default-memory now legitimately contributes one.
