+++ assistant/src/plugins/defaults/index.ts
@@ -64,7 +64,9 @@ import { resetMaxTokensContinueStoreForTests } from "./max-tokens-continue/conti
 import maxTokensContinuePostModelCall from "./max-tokens-continue/hooks/post-model-call.js";
 import maxTokensContinueStop from "./max-tokens-continue/hooks/stop.js";
 import maxTokensContinuePkg from "./max-tokens-continue/package.json" with { type: "json" };
+import memoryInit from "./memory/hooks/init.js";
 import memoryPostCompact from "./memory/hooks/post-compact.js";
+import memoryShutdown from "./memory/hooks/shutdown.js";
 import memoryUserPromptSubmit from "./memory/hooks/user-prompt-submit.js";
 import { memoryInjectors } from "./memory/injectors.js";
 import { memoryJobHandlers } from "./memory/job-handlers.js";
@@ -169,6 +171,8 @@ export const defaultMemoryPlugin: Plugin = {
     version: memoryPkg.version,
   },
   hooks: {
+    init: memoryInit,
+    shutdown: memoryShutdown,
     "user-prompt-submit": memoryUserPromptSubmit,
     "post-compact": memoryPostCompact,
   },
