+++ assistant/src/plugins/defaults/memory/filing-service.ts
@@ -1,17 +1,17 @@
 import { existsSync, readFileSync } from "node:fs";
 import { join } from "node:path";
 
-import { getConfig } from "../config/loader.js";
-import type { LLMCallSite } from "../config/schemas/llm.js";
+import { getConfig } from "../../../config/loader.js";
+import type { LLMCallSite } from "../../../config/schemas/llm.js";
 import {
   checkDiskPressureBackgroundGate,
   diskPressureBackgroundSkipLogFields,
   shouldLogDiskPressureBackgroundSkip,
-} from "../daemon/disk-pressure-background-gate.js";
-import { runBackgroundJob } from "../runtime/background-job-runner.js";
-import { getLogger } from "../util/logger.js";
-import { getWorkspaceDir } from "../util/platform.js";
-import { stripCommentLines } from "../util/strip-comment-lines.js";
+} from "../../../daemon/disk-pressure-background-gate.js";
+import { runBackgroundJob } from "../../../runtime/background-job-runner.js";
+import { getLogger } from "../../../util/logger.js";
+import { getWorkspaceDir } from "../../../util/platform.js";
+import { stripCommentLines } from "../../../util/strip-comment-lines.js";
 
 const log = getLogger("filing-service");
 
@@ -415,26 +415,6 @@ export class FilingService {
       );
     }
   }
-}
-
-/**
- * Construct and start the filing service singleton. Skipped under memory v2 —
- * the v2 consolidation job owns periodic background memory processing, so
- * running filing too would be redundant.
- */
-export function startFilingService(): void {
-  if (getConfig().memory.v2.enabled) {
-    log.info(
-      "Filing service skipped — memory v2 consolidation is the active background memory job",
-    );
-    return;
-  }
-  new FilingService().start();
-}
-
-/** Stop the filing service singleton if one is running; no-op otherwise. */
-export async function stopFilingService(): Promise<void> {
-  await FilingService.getInstance()?.stop();
 }
 
 function isWithinActiveHours(
